requirements.txt
# Ultra Doc-Intelligence Requirements

# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0

# ML/AI and embeddings
sentence-transformers>=3.0.0
torch>=2.1.0
numpy>=1.24.0

# HTTP requests for LLM API
requests==2.31.0
